+++ tests/testthat/test-spectrum.R
@@ -59,3 +59,15 @@ test_that("error thrown if zip contains non PJNZ files", {
   error <- expect_error(unroll_pjnz(file.path("refdata", "invalid_files.zip")))
   expect_equal(error$message, "Zip contains no PJNZ files")
 })
+
+
+test_that("extract_pjnz_naomi() returns country and region name", {
+
+  pjnz_zone <- system_file("extdata/demo-subnational-pjnz/demo_mwi2019_region-pjnz.zip")
+  spec <- extract_pjnz_naomi(pjnz_zone)
+
+  expect_setequal(spec$iso3, "MWI")
+  expect_setequal(spec$spectrum_country, "Malawi")
+  expect_setequal(spec$spectrum_region_code, c(10, 11, 12))
+  expect_setequal(spec$spectrum_region_name, c("Northern Region", "Central Region", "Southern Region"))
+})